requirements.txt
litellm
pyyaml
requests
beautifulsoup4
schedule
edge-tts
pygame
SpeechRecognition
pyaudio
groq